Move all the dashboard work under submission
