+++ opentech/apply/dashboard/urls.py
@@ -1,9 +1,8 @@
 from django.conf.urls import url
 
-from .views import DashboardView, SearchView
+from .views import DashboardView
 
 
 urlpatterns = [
     url(r'^$', DashboardView.as_view(), name="dashboard"),
-    url(r'^search$', SearchView.as_view(), name="search"),
 ]
