+++ opentech/apply/funds/templates/funds/includes/search.html
@@ -1,4 +1,4 @@
-<form action="{% url 'dashboard:search' %}" method="get" role="search" class="form form--header-search-desktop">
+<form action="{% url 'funds:search' %}" method="get" role="search" class="form form--header-search-desktop">
     <button class="button button--search" type="submit" aria-label="Search">
         <svg class="icon icon--magnifying-glass icon--search"><use xlink:href="#magnifying-glass"></use></svg>
     </button>
