+++ opentech/apply/funds/templates/funds/submissions_search.html
@@ -1,4 +1,4 @@
-{% extends "dashboard/dashboard.html" %}
+{% extends "funds/submissions.html" %}
 
 {% block page_header %}
     <div class="wrapper wrapper--medium">
